remove v1.History from serverless build
